Clean temp tex files

+++ cleaner.py
@@ -5,7 +5,7 @@ def clean():
     directory = os.listdir(".")
     print("Removing temporary files...")
     for item in directory:
-        if item.endswith((".txt", ".csv", ".jpg", ".png")) and item != "logo 512.png":
+        if item.endswith((".txt", ".csv", ".jpg", ".png", ".aux", ".fdb_latexmk", ".fls", ".log", ".out", ".synctex.gz", ".tex")) and item != "logo 512.png" and item != "latex.tex":
             os.remove(item)
